refactor: removed unused imports, moved to logical positions
